feat: implement company repositories for MongoDB

+++ R.Systems.Template.Core/Common/Lists/ListInfo.cs
@@ -3,6 +3,6 @@ namespace R.Systems.Template.Core.Common.Lists;
 public class ListInfo<T>
     where T : new()
 {
-    public int Count { get; init; }
+    public long Count { get; init; }
     public List<T> Data { get; init; } = new();
 }
